@@ -23,7 +23,7 @@
     "nanoid": "^5.0.3",
     "tsc": "^2.0.4",
     "tsx": "^3.13.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.2"
   },
   "devDependencies": {
     "@types/cookie-parser": "^1.4.6",
