@@ -1,4 +1,6 @@
-require('dotenv').config();
+import dotenv from 'dotenv';
+dotenv.config();
+
 import fs from 'fs';
 import crypto from 'node:crypto';
 
