@@ -1,4 +1,6 @@
-require("dotenv").config();
+import dotenv from 'dotenv';
+dotenv.config();
+
 import argon2 from 'argon2';
 import * as jose from 'jose'
 import crypto from 'node:crypto';
