@@ -1,10 +1,5 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-const mysql2_1 = __importDefault(require("mysql2"));
-const connectionPool = mysql2_1.default.createPool({
+import mysql from "mysql2";
+const connectionPool = mysql.createPool({
     host: process.env.HOST,
     user: process.env.DATABASE_USER,
     waitForConnections: true,
@@ -14,4 +9,4 @@ const connectionPool = mysql2_1.default.createPool({
     database: process.env.DATABASE_NAME,
     timezone: process.env.TIMEZONE,
 });
-exports.default = connectionPool;
+export default connectionPool;
