@@ -1,16 +1,34 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-require("dotenv").config();
-const express_1 = __importDefault(require("express"));
-const cors_1 = __importDefault(require("cors"));
-const server = (0, express_1.default)();
-server.use(cors_1.default);
-server.use(express_1.default.json());
-server.use(express_1.default.urlencoded({ extended: true }));
-let portNum = (Number(process.env.PORT) || 3000);
+import dotenv from 'dotenv';
+dotenv.config();
+import express from "express";
+import cors from 'cors';
+import cookieParser from 'cookie-parser';
+import mongoose from 'mongoose';
+//---
+import UserRouter from "./routers/user_router.js";
+import DeviceRouter from "./routers/device_router.js";
+//---
+const server = express();
+server.use(cors());
+server.use(cookieParser());
+server.use(express.json());
+server.use(express.urlencoded({ extended: true }));
+server.use("/api", UserRouter);
+server.use("/api", DeviceRouter);
+server.get('*', (request, response) => {
+    return response.status(404).json({ "error": "router not found" });
+});
+mongoose.connection.on("open", function () {
+    //console.log(ref);
+    console.log("Connected to mongo server.");
+});
+mongoose.connection.on("error", function (error) {
+    console.log("Could not connect to mongo server!");
+    console.log(error);
+});
+const uri = `mongodb+srv://${process.env.MONGO_USER}:${process.env.MONGO_PASS}@dev0.agidxfk.mongodb.net/?retryWrites=true&w=majority`;
+mongoose.connect(uri);
+let portNum = (Number(process.env.PORT));
 server.listen(portNum, () => {
     console.log("Server started on port " + portNum);
 });
